requirements.txt
# Testing and Quality
ruff
pytest

# Core Application Dependencies
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.1.0
google-genai==1.53.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
postgrest==2.25.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.3.2
psycopg-binary==3.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.2.1
realtime==2.25.0
rsa==4.9.1
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.0
websockets==15.0.1
Werkzeug==3.1.4
yarl==1.22.0
beautifulsoup4==4.12.3
requests==2.32.3

# Krag Bot Trading Dependencies
# MetaTrader5==5.0.45 ; platform_system == "Windows"
python-telegram-bot>=21.0
pandas>=2.2.0
numpy<2.0.0
PyYAML==6.0.1
